requirements.txt
streamlit
langchain
langchain-huggingface
langchain-community
google-search-results
langsmith
python-dotenv